закончен осталось по мелочи с пустыми строками
