+++ src/main/java/Main.java
@@ -40,7 +40,7 @@ public class Main
     // и игнорирующий введение любых других символов, строк и т.д.
     public static int waitInt (int numPeopleMin, int numPeopleMax)
     {
-        int numberPeople = 0;
+        int numberPeople;
         String message = "Введите целое число > %d и < %d";
         Scanner scanner = new Scanner(System.in);
         while (true)
@@ -67,76 +67,6 @@ public class Main
     public static void print(String massege)
     {
         System.out.println(massege);
-    }
-    // объявляем новый класс - строка меню(прайса)
-    public static class MenuBar
-    {
-        float priceDish;
-        String nameDish;
-
-        public MenuBar(String nameDish, float priceDish)
-        {
-            this.nameDish = nameDish;
-            this.priceDish = priceDish;
-        }
-        // далее метод определяющий окончание слова "рубль"
-        // в зависимости от полученной суммы товаров
-        public String printResult(float result)
-        {
-            String rubl = "Рублей";
-            if(result%100<=11&&result%100>=19)
-            {
-                rubl = "рублей";
-            }
-            switch(((int)result)%10)
-            {
-                case 0:
-                case 5:
-                case 6:
-                case 7:
-                case 8:
-                case 9:
-                    rubl = "рублей";
-                    break;
-                case 1:
-                    rubl = "рубль";
-                    break;
-                case 2:
-                case 3:
-                case 4:
-                    rubl = "рубля";
-                    break;
-            }
-            return rubl;
-        }
-        //далее метод заполняющий строку наименованиями
-        // введенных пользователем товаров
-        public String writeMenu()
-        {
-            Scanner scan = new Scanner(System.in);
-            nameDish = nameDish + "\n" + scan.nextLine();
-            return nameDish;
-        }
-        // далее метод запрашивающий цену товара и прибавляющий ее к
-        // сумме ранее введенных цен
-        public float priceSum()
-        {
-            Scanner scan = new Scanner(System.in);
-            while (true)
-            {
-                if (scan.hasNextDouble())
-                {
-                    priceDish = priceDish + scan.nextFloat();
-                    System.out.println("Товар успешно добавлен.\nДобавить еще товар? - введите любой символ!\n\nЧтобы завершить введите \"Завершить\"");
-                    return priceDish;
-                }
-                else
-                {
-                    System.out.println("Введите сумму товара в виде 00,00");
-                    scan.nextLine();
-                }
-            }
-        }
     }
 }
